web/requirements.txt
newspaper3k
lxml-html-clean